+++ R/align.R
@@ -2,7 +2,9 @@
 #'
 #' @param locked_package locked_package. In particular, the \code{version}
 #'    and \code{name} elements will be used.
-align <- function(locked_package) {
+#' @param verbose logical. Whether or not to announce usage of the package.
+#'    By default \code{TRUE}.
+align <- function(locked_package, verbose = TRUE) {
   if (is.list(locked_package) && !is.locked_package(locked_package)) {
     lapply(locked_package, align)
     ## Unlink the download directory to remove temporary files
@@ -17,6 +19,10 @@ align <- function(locked_package) {
 
   ## Symlink the locked package to the correct lockbox version.
   `symlink_to_lockbox!`(locked_package)
+
+  if (isTRUE(verbose)) {
+    announce_package_usage(locked_package$name, locked_package$version)
+  }
 }
 
 `symlink_to_lockbox!` <- function(locked_package) {